# src/multi-chat/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kuwa API",
        "description": "API definition for KuwaClient service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/create/bot": {
            "post": {
                "tags": [
                    "Bots"
                ],
                "summary": "Create a bot",
                "operationId": "f3e2fbb5bc1ec98f7cf05a51185c8f8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bot created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/read/bots": {
            "get": {
                "tags": [
                    "Bots"
                ],
                "summary": "List all bots available to the user",
                "operationId": "76609044293e45c254f20813f139af52",
                "responses": {
                    "200": {
                        "description": "Successful response returns a list of bots",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Indicates the status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "result": {
                                            "description": "An array of bot objects",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateBotRequest": {
                "properties": {
                    "llm_access_code": {
                        "type": "string"
                    },
                    "bot_name": {
                        "type": "string"
                    },
                    "visibility": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}